refactor: change paths to Provider animations and styles
